Add quotes around lines containing commas
Currently a response could be `{title: "Here's our 29 favourite tricks to thaw your cold cold heart."}` but you have a comma on that line which in CSV creates a new column. Surrounding the text with a quote keeps the line one column so the response will contain the comma again.